Adding Floating Sauce(No Texture)

+++ src/main/java/shit/randomfoodstuff/util/DurationHelper.java
@@ -6,16 +6,16 @@ public class DurationHelper {
 		return seconds * 20;
 	}
 	
-	public static int minutesToTicks(int minutes) {
-		return minutes * 1200;
+	public static int minutesToTicks(float minutes) {
+		return (int)(minutes * 1200);
 	}
 	
 	public static int ticksToSeconds(int ticks) {
 		return ticks / 20;
 	}
 	
-	public static int minutesToSeconds(int minutes) {
-		return minutes * 60;
+	public static int minutesToSeconds(float minutes) {
+		return (int)(minutes * 60);
 	}
 	
 }
